Add typescript config to the project
